pyproject.toml
[project]
name = "expense-tracker-backend"
version = "0.1.0"
description = "Expense Tracker API built with FastAPI and PostgreSQL"
authors = [{ name = "Jesreal Arcillas", email = "jesrealarcillas@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "fastapi (>=0.115.7,<0.116.0)",
  "uvicorn (>=0.34.0,<0.35.0)",
  "sqlalchemy (>=2.0.37,<3.0.0)",
  "psycopg2-binary (>=2.9.10,<3.0.0)",
  "pydantic[email] (>=2.10.6,<3.0.0)",
  "python-dotenv (>=1.0.1,<2.0.0)",
  "passlib[bcrypt] (>=1.7.4,<2.0.0)",
  "python-jose[cryptography] (>=3.3.0,<4.0.0)",
  "python-multipart (>=0.0.20,<0.0.21)",
]

[tool.poetry]
packages = [{ include = "expense_tracker_backend" }]

[project.scripts]
start = "uvicorn expense_tracker_backend.main:app --reload"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
